requirements.txt
langchain
python-dotenv
langchain_community
streamlit
langchain-groq
langchain_chroma
langchain_core
langchain_huggingface
langchain_text_splitters
chroma
pypdf
transformers
tiktoken
youtube_search
wikipedia
duckduckgo-search
xmltodict
certifi
chatgroq